Moved @tryghost/dynamic-routing-events to Ghost core
fix https://linear.app/ghost/issue/ENG-2065/dynamic-routing-events

- this package just exports one class, which is used by the server and
  frontend code, so I think it's OK to move it into shared until we have
  a better place to put it
- this then also removes the files for this package, as it's no longer
  needed
