+++ ghost/core/core/server/services/url/Resources.js
@@ -1,7 +1,7 @@
 const _ = require('lodash');
 const debug = require('@tryghost/debug')('services:url:resources');
 const DomainEvents = require('@tryghost/domain-events');
-const {URLResourceUpdatedEvent} = require('@tryghost/dynamic-routing-events');
+const URLResourceUpdatedEvent = require('../../../shared/events/URLResourceUpdatedEvent');
 const Resource = require('./Resource');
 const config = require('../../../shared/config');
 const models = require('../../models');
