+++ ghost/core/core/shared/events/URLResourceUpdatedEvent.js
