+++ ghost/core/package.json
@@ -84,7 +84,6 @@
     "@tryghost/debug": "0.1.32",
     "@tryghost/domain-events": "0.0.0",
     "@tryghost/donations": "0.0.0",
-    "@tryghost/dynamic-routing-events": "0.0.0",
     "@tryghost/email-analytics-provider-mailgun": "0.0.0",
     "@tryghost/email-analytics-service": "0.0.0",
     "@tryghost/email-content-generator": "0.0.0",
