--- ghost/dynamic-routing-events/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-    URLResourceUpdatedEvent: require('./lib/URLResourceUpdatedEvent')
-};
