--- ghost/dynamic-routing-events/test/dynamic-routing-events.test.js
@@ -1,8 +1,0 @@
-const assert = require('assert/strict');
-const events = require('../index');
-
-describe('Dynamic Routing Events', function () {
-    it('exports events', function () {
-        assert(events.URLResourceUpdatedEvent);
-    });
-});
